+++ src/shared/utils/types.js
@@ -4,16 +4,25 @@ import { isValid } from 'date-fns';
  * Returns true if value is of type number or a number as a string.
  */
 export function isNumber(value) {
+  if (value == null) return false;
+
+  // avoid dates
+  const t = typeof value;
+  if (t !== 'number' && t !== 'string') return false;
+
   const parsedValue = Number(value);
-  const isDate = value instanceof Date;
-  return value != null && !isDate && (parsedValue || parsedValue === 0);
+  return (parsedValue || parsedValue === 0);
 }
 
 /**
  * Returns true if value is a valid date instance.
  *
- * A date as a string will return false.
+ * A date as a string or a timestamp (number) will return false.
  */
 export function isValidDate(value) {
-  return value instanceof Date && isValid(value);
+  return (typeof value === 'object') && isValid(value);
+  // NOTEs:
+  // * Something like: `value instanceof Date` could be used,
+  //   but `global.Date` is mocked in tests, and this may evaluate to false
+  // * date-fns' `isValid` function returns true for a number
 }